Split test classes up and organised them a bit
