+++ Assembly-CSharp-Editor.csproj
@@ -46,11 +46,13 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-     <Compile Include="Assets\Editor\Tests\AgentUnitTests.cs" />
-     <Compile Include="Assets\Editor\Tests\GameManagerUnitTests.cs" />
-     <Compile Include="Assets\Editor\Tests\MapUnitTests.cs" />
-     <Compile Include="Assets\Editor\Tests\ResourceGroupUnitTests.cs" />
-     <Compile Include="Assets\Editor\Tests\RoboticonUnitTests.cs" />
+     <Compile Include="Assets\Editor\Tests\GameManagerTest.cs" />
+     <Compile Include="Assets\Editor\Tests\MapTest.cs" />
+     <Compile Include="Assets\Editor\Tests\MarketTest.cs" />
+     <Compile Include="Assets\Editor\Tests\PlayerTest.cs" />
+     <Compile Include="Assets\Editor\Tests\ResourceGroupTest.cs" />
+     <Compile Include="Assets\Editor\Tests\RoboticonTest.cs" />
+     <Compile Include="Assets\Editor\Tests\TileTest.cs" />
      <None Include="Assets\Standard Assets\Environment\Water\Water4\Shaders\FXWater4Simple.shader" />
      <None Include="Assets\Standard Assets\Environment\Water\Water\Shaders\FXWaterPro.shader" />
      <None Include="Assets\Resources\VFX\Shaders\GroundFromSpace.shader" />
