+++ Assets/Editor/Tests/GameManagerTest.cs
@@ -4,7 +4,7 @@ using NUnit.Framework;
 /// <summary>
 /// JBT created all new unit tests, as Bugfree chose to create their own testing framework instead of using an existing one
 /// </summary>
-public class GameManagerUnitTests
+public class GameManagerTest
 {
     [Test]
     public void GameWinTest()
