+++ Assets/Editor/Tests/MapTest.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 08f12fa5b91f7204f9709cb15da18194
+timeCreated: 1486441342
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
