+++ Assets/Editor/Tests/PlayerTest.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 5bf02967a72496d4e9ba92b83cde7400
-timeCreated: 1485127621
+guid: aeaa3d76e62b54e458467a9dc3bf256b
+timeCreated: 1486441244
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
