+++ Assets/Editor/Tests/RoboticonTest.cs
@@ -3,7 +3,7 @@
 /// <summary>
 /// JBT created all new unit tests, as Bugfree chose to create their own testing framework instead of using an existing one
 /// </summary>
-public class RoboticonUnitTests
+public class RoboticonTest
 {
     [Test]
     public void TestRoboticonName()
