+++ Assets/Editor/Tests/TileTest.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0f3db1c817f65b447bc07f94cf52f0a0
+timeCreated: 1486441252
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
